Refactor rows_between to be on the base_frame
